feat(query): add iceberg table functions
