+++ src/query/service/src/table_functions/table_function_factory.rs
@@ -32,6 +32,7 @@ use databend_common_storages_fuse::table_functions::FuseVacuumTemporaryTable;
 use databend_common_storages_fuse::table_functions::HilbertClusteringInfoFunc;
 use databend_common_storages_fuse::table_functions::SetCacheCapacity;
 use databend_common_storages_fuse::table_functions::TableFunctionTemplate;
+use databend_common_storages_iceberg::IcebergInspectTable;
 use databend_common_storages_stream::stream_status_table_func::StreamStatusTable;
 use databend_storages_common_table_meta::table_id_ranges::SYS_TBL_FUC_ID_END;
 use databend_storages_common_table_meta::table_id_ranges::SYS_TBL_FUNC_ID_BEGIN;
@@ -352,6 +353,16 @@ impl TableFunctionFactory {
         creators.insert(
             "show_roles".to_string(),
             (next_id(), Arc::new(ShowRoles::create)),
+        );
+
+        creators.insert(
+            "iceberg_snapshot".to_string(),
+            (next_id(), Arc::new(IcebergInspectTable::create)),
+        );
+
+        creators.insert(
+            "iceberg_manifest".to_string(),
+            (next_id(), Arc::new(IcebergInspectTable::create)),
         );
 
         TableFunctionFactory {
