+++ src/query/sql/src/planner/plans/scan.rs
@@ -236,8 +236,12 @@ impl Operator for Scan {
             }
             if let Some(col_stat) = v.clone() {
                 // Safe to unwrap: min, max are all `Some(_)`.
-                let min = col_stat.min.unwrap();
-                let max = col_stat.max.unwrap();
+                let Some(min) = col_stat.min.clone() else {
+                    continue;
+                };
+                let Some(max) = col_stat.max.clone() else {
+                    continue;
+                };
                 // ndv could be `None`, we will use `num_rows - null_count` as ndv instead.
                 //
                 // NOTE: don't touch the original num_rows, since it will be used in other places.
