+++ src/query/storages/fuse/src/table_functions/mod.rs
@@ -27,13 +27,12 @@ mod fuse_vacuum_drop_aggregating_index;
 mod fuse_vacuum_drop_inverted_index;
 mod fuse_vacuum_temporary_table;
 mod hilbert_clustering_information;
-mod table_args;
 
 mod set_cache_capacity;
 
 pub use clustering_information::ClusteringInformationFunc;
 pub use clustering_statistics::ClusteringStatisticsFunc;
-use databend_common_catalog::table_args::TableArgs;
+pub use databend_common_catalog::table_args::*;
 use databend_common_catalog::table_function::TableFunction;
 pub use function_template::SimpleTableFunc;
 pub use function_template::TableFunctionTemplate;
@@ -52,4 +51,3 @@ pub use fuse_vacuum_drop_inverted_index::FuseVacuumDropInvertedIndex;
 pub use fuse_vacuum_temporary_table::FuseVacuumTemporaryTable;
 pub use hilbert_clustering_information::HilbertClusteringInfoFunc;
 pub use set_cache_capacity::SetCacheCapacity;
-pub use table_args::*;
