+++ tests/sqllogictests/suites/tpch_iceberg/utils.test
@@ -54,6 +54,7 @@ partsupp
 region
 supplier
 
+
 ## test database
 statement ok
 create database abc;
